quic: submodules for connection
- Move substream-related logic into crate::connection::substream
- Rename/ move crate::upgrade -> crate::connection::connecting
